requirements.txt
fastapi==0.136.0
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
langchain==0.3.27
langchain-community==0.3.27
langchain-chroma==0.2.6
langchain-google-genai==2.0.11
chromadb==1.5.8
pypdf==5.1.0
sentence-transformers==3.4.1
torch==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu